requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
decorator==5.2.1
Django==5.2.3
django-celery-beat==2.8.1
django-timezone-field==7.1
django_celery_results==2.6.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.14
h11==0.16.0
idna==3.10
ipython==8.37.0
jedi==0.19.2
kombu==5.5.4
matplotlib-inline==0.1.7
nsepython==2.97
numpy==2.2.6
packaging==25.0
pandas==2.3.0
parso==0.8.4
prompt_toolkit==3.0.51
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.2.0
requests==2.32.4
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
vine==5.1.0
wcwidth==0.2.13
